Create a feature flag helper for tests.
